+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When evacuating or migrating an InUse resources, wait until the resource is no longer in-use to remove resource.
 - Restoring a pre-SP-mixing-backup into an SP mixing scenario
 - SpMixing: ZFS <-> ZFS_THIN is no longer considered as mixed SP
+- Encryption: don't create error-reports for user errors
 
 ## [1.26.0] - 2024-01-29